updated helper method
